refactor(ricer::config): move toml parsing and tests into separate file
