+++ src/config.rs
@@ -26,13 +26,15 @@
 //! [`DefaultLocator`]: crate::locate::DefaultLocator
 
 mod settings;
+mod toml;
 
 #[doc(inline)]
 pub use settings::*;
+pub use toml::*;
 
 use crate::locate::Locator;
 
-use log::{debug, info, trace};
+use log::debug;
 use mkdirp::mkdirp;
 use std::{
     fmt,
@@ -40,9 +42,7 @@ use std::{
     io,
     io::{Read, Write},
     path::{Path, PathBuf},
-    str::FromStr,
 };
-use toml_edit::{DocumentMut, Item, Key, Table};
 
 /// Error types for [`ConfigFile`].
 #[derive(Debug, thiserror::Error)]
@@ -61,22 +61,6 @@ pub enum ConfigFileError {
 
     #[error("Failed to parse '{path}'")]
     Toml { source: TomlError, path: PathBuf },
-}
-
-/// Error types for [`Toml`].
-#[derive(Debug, PartialEq, Eq, thiserror::Error)]
-pub enum TomlError {
-    #[error("Failed to parse TOML data")]
-    BadParse { source: toml_edit::TomlError },
-
-    #[error("TOML table '{table}' not found")]
-    TableNotFound { table: String },
-
-    #[error("TOML table '{table}' not defined as a table")]
-    NotTable { table: String },
-
-    #[error("TOML entry '{key}' not found in table '{table}'")]
-    EntryNotFound { table: String, key: String },
 }
 
 /// Format preserving configuration file handler.
@@ -259,215 +243,6 @@ where
     }
 }
 
-/// TOML parser.
-///
-/// Offers basic CRUD interface for TOML parsing. Expects TOML data in string
-/// form. Leaves file handling to caller. Mainly operates on whole tables for
-/// key-value pair manipulation. Note, that `document` is terminology used to
-/// refer to parsed TOML data.
-///
-/// # Invariants
-///
-/// 1. Preserve original formatting of document.
-///
-/// # See also
-///
-/// - [`ConfigFile`]
-#[derive(Clone, Default, Debug)]
-pub struct Toml {
-    doc: DocumentMut,
-}
-
-impl Toml {
-    pub fn new() -> Self {
-        trace!("Construct new TOML parser");
-        Self { doc: DocumentMut::new() }
-    }
-
-    /// Add TOML entry into document.
-    ///
-    /// Will add given `entry` into target `table`. If `table` does not exist, then it
-    /// will be created and `entry` will be inserted into it.
-    ///
-    /// Will replace any entries that match the key in `entry`, returning the
-    /// old entry that was replaced. If no replacement took place, then `None`
-    /// is returned instead.
-    ///
-    /// # Errors
-    ///
-    /// - Return [`TomlError::NotTable`] if target table was not defined as
-    ///   a table.
-    ///
-    /// [`TomlError::NotTable`]: crate::config::TomlError::NotTable
-    pub fn add(
-        &mut self,
-        table: impl AsRef<str>,
-        entry: (Key, Item),
-    ) -> Result<Option<(Key, Item)>, TomlError> {
-        let (key, value) = entry;
-        info!("Add TOML entry '{}' to '{}' table", key.get(), table.as_ref());
-        let entry = match self.get_table_mut(table.as_ref()) {
-            Ok(table) => table,
-            Err(TomlError::TableNotFound { .. }) => {
-                let mut new_table = Table::new();
-                new_table.set_implicit(true);
-                self.doc.insert(table.as_ref(), Item::Table(new_table));
-                self.doc[table.as_ref()].as_table_mut().unwrap()
-            }
-            Err(err) => return Err(err),
-        };
-        let entry = entry.insert(key.get(), value).map(|old| (key, old));
-        Ok(entry)
-    }
-
-    /// Get entry from target table in document.
-    ///
-    /// Return reference to full key-value pair in document.
-    ///
-    /// # Errors
-    ///
-    /// - Return [`TomlError::TableNotFound`] if target table is not found
-    ///   in document.
-    /// - Return [`TomlError::NotTable`] if target table was not defined as
-    ///   a table.
-    /// - Return [`TomlError::EntryNotFound`] if target key-value pair
-    ///   is not found in document.
-    ///
-    /// [`TomlError::TableNotFound`]: crate::config::TomlError::TableNotFound
-    /// [`TomlError::NotTable`]: crate::config::TomlError::NotTable
-    /// [`TomlError::EntryNotFound`]: crate::config::TomlError::EntryNotFound
-    pub fn get<S>(&self, table: S, key: S) -> Result<(&Key, &Item), TomlError>
-    where
-        S: AsRef<str>,
-    {
-        info!("Get TOML entry '{}' from '{}' table", key.as_ref(), table.as_ref());
-        let entry = self.get_table(table.as_ref())?;
-        let entry = entry.get_key_value(key.as_ref()).ok_or_else(|| TomlError::EntryNotFound {
-            table: table.as_ref().into(),
-            key: key.as_ref().into(),
-        })?;
-        Ok(entry)
-    }
-
-    /// Rename TOML entry from document.
-    ///
-    /// Rename entry from target `table`. Returns old unrenamed entry.
-    ///
-    /// # Errors
-    ///
-    /// - Return [`TomlError::TableNotFound`] if target table is not found
-    ///   in document.
-    /// - Return [`TomlError::NotTable`] if target table was not defined as
-    ///   a table.
-    /// - Return [`TomlError::EntryNotFound`] if target key-value pair
-    ///   is not found in document.
-    ///
-    /// [`TomlError::TableNotFound`]: crate::config::TomlError::TableNotFound
-    /// [`TomlError::NotTable`]: crate::config::TomlError::NotTable
-    /// [`TomlError::EntryNotFound`]: crate::config::TomlError::EntryNotFound
-    pub fn rename<S>(&mut self, table: S, from: S, to: S) -> Result<(Key, Item), TomlError>
-    where
-        S: AsRef<str>,
-    {
-        let entry = self.get_table_mut(table.as_ref())?;
-        let (old_key, old_item) = entry.remove_entry(from.as_ref()).ok_or_else(|| {
-            TomlError::EntryNotFound { table: table.as_ref().into(), key: from.as_ref().into() }
-        })?;
-
-        // INVARIANT: preserve original formatting that existed beforehand.
-        let new_key = Key::new(to.as_ref()).with_leaf_decor(old_key.leaf_decor().clone());
-        entry.insert_formatted(&new_key, old_item.clone());
-
-        Ok((old_key, old_item))
-    }
-
-    /// Remove TOML entry from document.
-    ///
-    /// Remove `key` from target `table`. Returns removed entry.
-    ///
-    /// # Errors
-    ///
-    /// - Return [`TomlError::TableNotFound`] if target table is not found
-    ///   in document.
-    /// - Return [`TomlError::NotTable`] if target table was not defined as
-    ///   a table.
-    /// - Return [`TomlError::EntryNotFound`] if target key-value pair
-    ///   is not found in document.
-    ///
-    /// [`TomlError::TableNotFound`]: crate::config::TomlError::TableNotFound
-    /// [`TomlError::NotTable`]: crate::config::TomlError::NotTable
-    /// [`TomlError::EntryNotFound`]: crate::config::TomlError::EntryNotFound
-    pub fn remove<S>(&mut self, table: S, key: S) -> Result<(Key, Item), TomlError>
-    where
-        S: AsRef<str>,
-    {
-        let entry = self.get_table_mut(table.as_ref())?;
-        let entry = entry.remove_entry(key.as_ref()).ok_or_else(|| TomlError::EntryNotFound {
-            table: table.as_ref().into(),
-            key: key.as_ref().into(),
-        })?;
-        Ok(entry)
-    }
-
-    /// Get target table in document.
-    ///
-    /// Return reference to target table in document.
-    ///
-    /// # Errors
-    ///
-    /// - Return [`TomlError::TableNotFound`] if target table is not found
-    ///   in document.
-    /// - Return [`TomlError::NotTable`] if target table was not defined as
-    ///   a table.
-    ///
-    /// [`TomlError::TableNotFound`]: crate::config::TomlError::TableNotFound
-    /// [`TomlError::NotTable`]: crate::config::TomlError::NotTable
-    pub(crate) fn get_table(&self, key: &str) -> Result<&Table, TomlError> {
-        debug!("Get TOML table '{key}'");
-        let table =
-            self.doc.get(key).ok_or_else(|| TomlError::TableNotFound { table: key.into() })?;
-        let table = table.as_table().ok_or_else(|| TomlError::NotTable { table: key.into() })?;
-        Ok(table)
-    }
-
-    /// Get mutable target table in document.
-    ///
-    /// Return mutable reference to target table in document.
-    ///
-    /// # Errors
-    ///
-    /// - Return [`TomlError::TableNotFound`] if target table is not found
-    ///   in document.
-    /// - Return [`TomlError::NotTable`] if target table was not defined as
-    ///   a table.
-    ///
-    /// [`TomlError::TableNotFound`]: crate::config::TomlError::TableNotFound
-    /// [`TomlError::NotTable`]: crate::config::TomlError::NotTable
-    pub(crate) fn get_table_mut(&mut self, key: &str) -> Result<&mut Table, TomlError> {
-        debug!("Get mutable TOML table '{key}'");
-        let table =
-            self.doc.get_mut(key).ok_or_else(|| TomlError::TableNotFound { table: key.into() })?;
-        let table =
-            table.as_table_mut().ok_or_else(|| TomlError::NotTable { table: key.into() })?;
-        Ok(table)
-    }
-}
-
-impl fmt::Display for Toml {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.doc)
-    }
-}
-
-impl FromStr for Toml {
-    type Err = TomlError;
-
-    fn from_str(data: &str) -> Result<Self, Self::Err> {
-        let doc: DocumentMut = data.parse().map_err(|err| TomlError::BadParse { source: err })?;
-        Ok(Self { doc })
-    }
-}
-
 /// TOML serialization and deserialization configuration.
 ///
 /// Interface to simplify serialization and deserialization of parsed TOML data.
@@ -574,6 +349,3 @@ impl Config for CmdHookConfig {
         locator.hooks_config()
     }
 }
-
-#[cfg(test)]
-mod tests {}
